New import statement seems to have fixed one of our biggest bugs:
If a user was signed in and then the server restarted, trying to go
directly to the home.jsp page after the restart would cause the site to
crash. 

This bug appeared to be due to corruption of the HttpSession. The new
import statement ensures that the HttpSession attributes user, username,
and userid are always populated by refreshing the user's state on every
page load.

I went ahead and added the new import statement to several of the main
JSP's on our site.