Welcome page actually working now. RootServlet was a very bad idea.
Our homepage is now available at:
localhost:8080/QuizSite/

You no longer need to include "home.jsp" or "welcomepage.jsp".

My first strategy to make it do this was to create a RootServlet and map
it to the URL "/". But doing this actually causes all URLs to be passed
to the RootServlet, meaning the entire website becomes unavailable.

The correct implementation is through web.xml. Web.xml specifies the
names of "welcome files", which are the default files it loads when none
is specified. Since index.jsp is in the list, I created an index.jsp
file that simply redirects to home.jsp. Problem solved!